+++ src/rest-server/src/utils/frameworkConverter.js
@@ -1,4 +1,3 @@
-const base32 = require('base32');
 const zlib = require('zlib');
 const _ = require('lodash');
 const axios = require('axios');
@@ -9,21 +8,6 @@ const logger = require('@pai/config/logger');
 const launcherConfig = require('@pai/config/launcher');
 const {apiserver} = require('@pai/config/kubernetes');
 const k8s = require('@pai/utils/k8sUtils');
-
-const convertName = (name) => {
-  // convert framework name to fit framework controller spec
-  return name.toLowerCase().replace(/[^a-z0-9]/g, '');
-};
-
-const encodeName = (name) => {
-  if (name.startsWith('unknown') || !name.includes('~')) {
-    // framework is not generated by PAI
-    return convertName(name.replace(/^unknown/g, ''));
-  } else {
-    // base32 encode
-    return base32.encode(name);
-  }
-};
 
 const decodeName = (name, labels) => {
   if (labels && labels.jobName) {
@@ -42,7 +26,7 @@ const decompressField = (val) => {
   }
 };
 
-const convertState = (state, exitCode, retryStatus) => {
+const convertState = (state, exitCode) => {
   switch (state) {
     case 'AttemptCreationPending':
     case 'AttemptCreationRequested':
@@ -74,7 +58,7 @@ const convertState = (state, exitCode, retryStatus) => {
   }
 };
 
-const convertToJobRetry = (framework) => {
+const convertToJobAttempt = (framework) => {
   const completionStatus = framework.status.attemptStatus.completionStatus;
   const name = decodeName(framework.metadata.name, framework.metadata.labels);
   const frameworkName = framework.metadata.name;
@@ -87,14 +71,14 @@ const convertToJobRetry = (framework) => {
     framework.status.retryPolicyStatus.retryDelaySec,
   );
   const originState = framework.status.state;
-  const retryIndex = framework.status.attemptStatus.id;
+  const attemptIndex = framework.status.attemptStatus.id;
   const jobStartedTime = new Date(
     framework.metadata.creationTimestamp,
   ).getTime();
-  const retryStartedTime = new Date(
+  const attemptStartedTime = new Date(
     framework.status.attemptStatus.startTime,
   ).getTime();
-  const retryCompletedTime = new Date(
+  const attemptCompletedTime = new Date(
     framework.status.attemptStatus.completionTime,
   ).getTime();
   const totalGpuNumber = framework.metadata.annotations
@@ -111,17 +95,17 @@ const convertToJobRetry = (framework) => {
     username,
     state,
     originState,
-    retryIndex,
+    attemptIndex,
     jobStartedTime,
-    retryStartedTime,
-    retryCompletedTime,
+    attemptStartedTime,
+    attemptCompletedTime,
     totalGpuNumber,
     totalTaskNumber,
     totalTaskRoleNumber,
   };
 };
 
-const convertToJobRetryDetail = async (framework, retryIndex) => {
+const convertToJobAttemptDetail = async (framework, attemptIndex) => {
   // check fields which may be compressed
   if (framework.status.attemptStatus.taskRoleStatuses == null) {
     framework.status.attemptStatus.taskRoleStatuses = decompressField(
@@ -349,6 +333,6 @@ const extractRuntimeOutput = (podCompletionStatus) => {
 
 // module exports
 module.exports = {
-  convertToJobRetry,
-  convertToJobRetryDetail,
+  convertToJobAttempt,
+  convertToJobAttemptDetail,
 };
